requirements.txt
mysql-connector-python
eto
